+++ src/main/clojure/com/github/clojure_repl/intellij/action/eval.clj
@@ -1,6 +1,7 @@
 (ns com.github.clojure-repl.intellij.action.eval
   (:require
    [clojure.string :as string]
+   [com.github.clojure-repl.intellij.actions :as actions]
    [com.github.clojure-repl.intellij.db :as db]
    [com.github.clojure-repl.intellij.nrepl :as nrepl]
    [com.github.clojure-repl.intellij.parser :as parser]
@@ -15,14 +16,12 @@
    [com.intellij.openapi.actionSystem CommonDataKeys]
    [com.intellij.openapi.actionSystem AnActionEvent]
    [com.intellij.openapi.editor Editor]
-   [com.intellij.openapi.project Project]
    [com.intellij.openapi.vfs VirtualFile]))
 
 (set! *warn-on-reflection* true)
 
 (defn ^:private eval-action [^AnActionEvent event loading-msg eval-fn success-msg-fn {:keys [inlay-hint-feedback?]}]
   (when-let [editor ^Editor (.getData event CommonDataKeys/EDITOR_EVEN_IF_INACTIVE)]
-
     (let [project (.getProject editor)]
       (if (db/get-in project [:current-nrepl :session-id])
         (tasks/run-background-task!
@@ -74,6 +73,11 @@
          (string/join "\n" (:value response)))
        {:inlay-hint-feedback? true}))))
 
+(defn interrupt [^AnActionEvent event]
+  (-> event
+      actions/action-event->project
+      nrepl/interrupt))
+
 (defn eval-defun-action [^AnActionEvent event]
   (eval-action
    event
@@ -92,22 +96,18 @@
    {:inlay-hint-feedback? true}))
 
 (defn clear-repl-output-action [^AnActionEvent event]
-  (let [editor ^Editor (.getData event CommonDataKeys/EDITOR_EVEN_IF_INACTIVE)
-        project ^Project (or (.getData event CommonDataKeys/PROJECT)
-                             (.getProject editor))]
+  (let [project (actions/action-event->project event)]
     (ui.repl/clear-repl project (db/get-in project [:console :ui]))))
 
 (defn history-up-action [^AnActionEvent event]
-  (let [editor ^Editor (.getData event CommonDataKeys/EDITOR_EVEN_IF_INACTIVE)
-        project ^Project (or (.getData event CommonDataKeys/PROJECT)
-                             (.getProject editor))]
-    (ui.repl/history-up project)))
+  (-> event
+      actions/action-event->project
+      ui.repl/history-up))
 
 (defn history-down-action [^AnActionEvent event]
-  (let [editor ^Editor (.getData event CommonDataKeys/EDITOR_EVEN_IF_INACTIVE)
-        project ^Project (or (.getData event CommonDataKeys/PROJECT)
-                             (.getProject editor))]
-    (ui.repl/history-down project)))
+  (-> event
+      actions/action-event->project
+      ui.repl/history-down))
 
 (defn switch-ns-action [^AnActionEvent event]
   (eval-action